+++ source/smolv.h
@@ -23,7 +23,7 @@
 // - Reordering instruction opcodes so that the most common ones are the smallest values, for smaller
 //  varint encoding.
 // - Encoding several instructions in a more compact form, e.g. the "typical <=4 component swizzle"
-//  shape of a VectorShuffle instruction.
+//  shape of a VectorShuffle instruction, or sequences of MemberDecorate instructions.
 //
 // A somewhat similar utility is spirv-remap from glslang, see
 // https://github.com/KhronosGroup/glslang/blob/master/README-spirv-remap.txt